Update migrations and jo

+++ database/migrations/2018_09_21_082032_create_job_orders_table.php
@@ -17,7 +17,7 @@ class CreateJobOrdersTable extends Migration
             $table->increments('id');
             $table->integer('inspection_id')->unsigned();
             $table->date('start');
-            $table->integer('progressCount')->default(0);
+            $table->float('progressCount',5,2)->default(0);
             $table->boolean('isStatus')->default(0);
             $table->datetime('end')->nullable();
             $table->text('remarks')->nullable();
